ACCUMULO-3599 Get (un)secure ITs running against mini/standalone
A large series of changes to get all of the ITs running a variety of ways.
By default, ITs still run with unsecure and miniclusters. Kerberos can
be enabled with MINI. Additionally, all ITs should run against an unsecure
normal installation as well as a normal installation with Kerberos on.

* Push down Hadoop configs to PrintInfo
* Support users and keytabs (or passwords) in cluster configuration
* Use the filesystem from the cluster
* Reduce Hadoop UGI spam in the logs
* Ensure that the log reference seen is for the table we made, not the trace table.
* Get ProxyITs passing regardless of client configuration installed.
* Use a CountDownLatch to coordinate scanners running on ScanIdIT
* Move UserCompactionStrategyIT strategies into test jar to enable standalone tests.
* Apply extra quoting to args for standalone cluster exec.