Initial commit of the work to stand up an accumulo-proxy inside a docker container.
I have only implemented support for Accumulo 2.x and by default this first commit contains:
* Accumulo 2.0.0
* Hadoop 3.2.1
* Zookeeper 3.5.7

A new document (DOCKER.md) has been created to start to document the implementation and usage guide which should allow others to test this if they so wished.

A number of outstanding questions will be posted on the issue, there is also a number of TODOs still required to be implemented that I'm tracking.