Redirect cluster logs to files to reduce output
When the tests are run, the exception messages from the proxy server
are sent to STDOUT. This pollutes the output and makes it harder to
read the test output.

This change redirects the server messages to a file and keeps the output
clean.

Also replaced a deprecated property with its newer version to remove a
build warning message.