Collapsed tests to speed up runtime.
These single tests all trigger the BeforeEach method which
creates new clients, tables, and namespaces. They also trigger the
AfterEach method which deletes the table and namespace and also closes
the client.

This commit collapses the invalid auth tests into a single test case so
the table creation and deletion is not triggered for each auth attempt.